refactor(Visualize_RD.py): change visualize_simulation function signature and add support for loading substrate data directly as a NumPy array
The visualize_simulation function in Visualize_RD.py has been refactored to accept either the path to a substrate data file or the substrate data directly as a NumPy array. Previously, it only accepted the path to a substrate data file.

The function signature has been changed to:
```
def visualize_simulation(substrate_source, plot_title="", rate=4, envelop=False, show_average=False):
```

The `substrate_source` parameter can now be either a string specifying the path to the substrate data file or the substrate data itself as a NumPy array.

If `substrate_source` is a string, the function loads the substrate data from the file using `np.loadtxt()`. If `substrate_source` is a NumPy array, it is directly used as the substrate data.

This change improves the flexibility of the function and